REM: unneeded old data functions
