+++ src/GasModels.jl
@@ -43,8 +43,7 @@ module GasModels
     include("core/base.jl")
     include("core/types.jl")
     include("core/unit_converters.jl")
-    include("core/data_new.jl")
-    # include("core/data.jl")  # disable to find still needed functions
+    include("core/data.jl")
     include("core/variable.jl")
     include("core/constraint.jl")
     include("core/constraint_template.jl")
